perf: Revamp error handler
